feat(OwnersInfo): move dv-object to core domain
