+++ src/core/domain/models/DvObjectOwnerNode.ts
