+++ src/core/index.ts
@@ -1,3 +1,4 @@
 export { ReadError } from './domain/repositories/ReadError';
 export { WriteError } from './domain/repositories/WriteError';
 export { ApiConfig } from './infra/repositories/ApiConfig';
+export { DvObjectOwnerNode } from './domain/models/DvObjectOwnerNode';
