+++ src/core/infra/repositories/transformers/OwnerNodePayload.ts
