--- src/dv-object/index.ts
@@ -1,1 +1,0 @@
-export { DvObjectOwnerNode } from './domain/models/DvObjectOwnerNode';
