+++ src/index.ts
@@ -5,4 +5,3 @@ export * from './auth';
 export * from './datasets';
 export * from './metadataBlocks';
 export * from './files';
-export * from './dv-object';
