+++ test/testHelpers/datasets/datasetHelper.ts
@@ -3,7 +3,7 @@ import { NodeHtmlMarkdown } from 'node-html-markdown';
 import axios, { AxiosResponse } from 'axios';
 import { TestConstants } from '../TestConstants';
 import { DatasetPayload } from '../../../src/datasets/infra/repositories/transformers/DatasetPayload';
-import { DvObjectType } from '../../../src/dv-object/domain/models/DvObjectOwnerNode';
+import { DvObjectType } from '../../../src/core/domain/models/DvObjectOwnerNode';
 
 const DATASET_CREATE_TIME_STR = '2023-05-15T08:21:01Z';
 const DATASET_UPDATE_TIME_STR = '2023-05-15T08:21:03Z';
